requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
prettyprint==0.1.5
pytz==2019.3
requests2==2.16.0
six==1.13.0
soupsieve==1.9.5
tzlocal==2.0.0
urllib3==1.21.1
uWSGI==2.0.18
Werkzeug==0.16.0
